refactor: move ports to contracts
