--- internal/core/ports/doc.go
@@ -1,2 +1,0 @@
-// Package ports contains exposed ports.
-package ports
