+++ internal/provider/publiccloud/contracts/doc.go
@@ -1,0 +1,2 @@
+// Package contracts contains service & repository interfaces.
+package contracts
