+++ internal/provider/publiccloud/contracts/public_cloud_repository.go
@@ -1,4 +1,4 @@
-package ports
+package contracts
 
 import (
 	"context"
